moved api/client/ to api/ removed api/admin.
