+++ ddpui/tests/api_tests/test_warehouse_api.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock, patch
 
-from ddpui.api.client.warehouse_api import (
+from ddpui.api.warehouse_api import (
     get_schema,
     get_table,
     get_table_columns,
@@ -9,7 +9,7 @@ from ddpui.api.client.warehouse_api import (
 
 
 @patch.multiple(
-    "ddpui.api.client.warehouse_api",
+    "ddpui.api.warehouse_api",
     get_warehouse_data=Mock(return_value=["table1", "table2"]),
 )
 def test_get_table_success():
@@ -23,7 +23,7 @@ def test_get_table_success():
 
 
 @patch.multiple(
-    "ddpui.api.client.warehouse_api",
+    "ddpui.api.warehouse_api",
     get_warehouse_data=Mock(return_value=["schema1", "schema2"]),
 )
 def test_get_schema_success():
@@ -36,7 +36,7 @@ def test_get_schema_success():
 
 
 @patch.multiple(
-    "ddpui.api.client.warehouse_api",
+    "ddpui.api.warehouse_api",
     get_warehouse_data=Mock(return_value=["column1", "column2"]),
 )
 def test_get_table_columns_success():
@@ -51,7 +51,7 @@ def test_get_table_columns_success():
 
 
 @patch.multiple(
-    "ddpui.api.client.warehouse_api",
+    "ddpui.api.warehouse_api",
     get_warehouse_data=Mock(
         return_value=[{"column_1": "value_1"}, {"column2": "value2}"}]
     ),
