+++ ddpui/tests/api_tests/test_webhook_api.py
@@ -10,7 +10,7 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ddpui.settings")
 os.environ["DJANGO_ALLOW_ASYNC_UNSAFE"] = "true"
 django.setup()
 
-from ddpui.api.client.webhook_api import (
+from ddpui.api.webhook_api import (
     FLOW_RUN,
     get_message_type,
     get_flowrun_id_and_state,
